feat(location): implémente le reverse geocoding et améliore la précision des coordonnées - Crée une branche `reverseGeocoding` pour gérer le reverse geocoding des `Location`. - Ajoute les attributs `country`, `city` et `countryCode` au BO `Location`. - Implémente un `ReverseGeocodingService` pour interagir avec l'API Nominatim via la méthode `reverseGeocode`. - Crée un DTO `GeocodingResult` pour structurer les échanges entre `LocationService` et `ReverseGeocodingService`. - Améliore la précision des coordonnées :   - Passe les attributs `latitude` et `longitude` de `double` à `BigDecimal` dans le BO `Location`.   - Remplace les colonnes `FLOAT` par `DECIMAL(9,6)` en base de données.

+++ workspace/springboot-app/src/main/java/fr/memoires_vives/bll/LocationServiceImpl.java
@@ -1,5 +1,6 @@
 package fr.memoires_vives.bll;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -40,26 +41,27 @@ public class LocationServiceImpl implements LocationService {
 			east = 180;
 			west = -180;
 		}
-		
+
 		while (west < -180) {
 			west += 360;
 		}
 		while (east > 180) {
 			east -= 360;
 		}
-		return locationRepository.findInSquare(north, south, east, west);
+		return locationRepository.findInSquare(BigDecimal.valueOf(north), BigDecimal.valueOf(south),
+				BigDecimal.valueOf(east), BigDecimal.valueOf(west));
 	}
 
 	@Override
 	public Location createFromCoordinates(double latitude, double longitude) {
 		Location location = new Location();
 		location.setName(createLocationName());
-		location.setLatitude(latitude);
-		location.setLongitude(longitude);
-		
+		location.setLatitude(BigDecimal.valueOf(latitude));
+		location.setLongitude(BigDecimal.valueOf(longitude));
+
 		return location;
 	}
-	
+
 	private String createLocationName() {
 		return UUID.randomUUID().toString().substring(0, 30);
 	}
